Fix bignum cross-allocation corruption: int vs long s_dot mismatch
The local s_dot in linux_x86_64/bignum.c declared I as `long`, but the
kernel-wide `struct sdot` in dfuncs.h declares I as `int`. On i386
long==int so this coincidence held; on x86_64 LP64 reading p->I picks
up 4 bytes of struct padding into the high half, and carry >>= 30 in
adbig's add loop shifts that garbage down into the low 30-bit window
of the next limb -- corrupting every bignum result with more than one
limb.

Test suite expanded with bignum equality cases; KNOWN-BUGS.md cleared
of the three fixed entries.

+++ franz/linux_x86_64/bignum.c
@@ -343,8 +343,16 @@ long mul, add;
     }
 }
 
-/* adbig -- bignum addition via stack-allocated limb arrays. */
-struct s_dot { long I; struct s_dot *CDR; };
+/* adbig -- bignum addition via stack-allocated limb arrays.
+ *
+ * IMPORTANT: this local s_dot must mirror the kernel's `struct sdot`
+ * (defined in dfuncs.h: `int I; lispval CDR;`). On i386 long==int, but
+ * on x86_64 LP64 long is 8 bytes -- using `long I` here would cause
+ * `p->I` to read 4 bytes of struct padding into the high half of the
+ * value, which then gets shifted into the low limbs by `carry >>= 30`
+ * and corrupts every limb above the lowest one.
+ */
+struct s_dot { int I; struct s_dot *CDR; };
 extern struct s_dot *export();
 
 struct s_dot *
